pyproject.toml
[project]
name = "chess-analyse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "plotly>=6.5.2",
    "python-chess>=1.999",
    "tqdm>=4.67.1",
]
